refactor: rename SimplifiedDI to Dependencies
Remove evidence of past over-engineering by using a clean,
straightforward name that describes what the file does.
